Improved shutdown behaviour to properly handle client mode (don't delete server socket if you are a client).  Also added main MDP ping loop of sending packets.  Reading returned packets not yet implemented.  Server handling of MDP_TX packets also not yet implemented.

+++ dna.c
@@ -305,8 +305,12 @@ void servalShutdownCleanly()
   unlink(filename);
   snprintf(filename,1024,"%s/serval.pid",instancepath);
   unlink(filename);
-  snprintf(filename,1024,"%s/mdp.socket",instancepath);
-  unlink(filename);
+  if (mdp_client_socket==-1) {
+    snprintf(filename,1024,"%s/mdp.socket",instancepath);
+    unlink(filename);
+  } else {
+    overlay_mdp_client_done();
+  }
   exit(0);
 }
